requirements.txt
art==5.7
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-discover-runner==1.0
django-filter==22.1
django-role-permissions==3.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
fish-text-ru==2.0.2
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
names==0.3.0
Pillow==9.2.0
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
zipp==3.8.1
